Merge feature with master

+++ app/src/main/java/com/coders/TaskApp/Utils/TodoTouchHelperCallback.java
@@ -1,4 +1,4 @@
-package com.coders.TaskApp;
+package com.coders.TaskApp.Utils;
 
 import android.content.Context;
 import android.graphics.Canvas;
@@ -12,17 +12,21 @@ import androidx.core.content.ContextCompat;
 import androidx.recyclerview.widget.ItemTouchHelper;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.coders.TaskApp.Adapter.TodoAdapter;
+import com.coders.TaskApp.R;
+import com.coders.TaskApp.Repository.TaskRepository;
 import com.coders.TaskApp.models.Todo;
+import com.google.android.material.snackbar.BaseTransientBottomBar;
 import com.google.android.material.snackbar.Snackbar;
 
 public class TodoTouchHelperCallback extends ItemTouchHelper.SimpleCallback {
 
-    CustomAdapter adapter;
+    TodoAdapter adapter;
     Context context;
     View parent;
     TaskRepository repository;
 
-    public TodoTouchHelperCallback(int dragDirs, int swipeDirs, CustomAdapter adapter, Context context, View parent) {
+    public TodoTouchHelperCallback(int dragDirs, int swipeDirs, TodoAdapter adapter, Context context, View parent) {
         super(dragDirs, swipeDirs);
         this.adapter = adapter;
         this.context = context;
@@ -37,17 +41,13 @@ public class TodoTouchHelperCallback extends ItemTouchHelper.SimpleCallback {
 
     @Override
     public void onChildDraw(@NonNull Canvas c, @NonNull RecyclerView recyclerView, @NonNull RecyclerView.ViewHolder viewHolder, float dX, float dY, int actionState, boolean isCurrentlyActive) {
-        super.onChildDraw(c, recyclerView, viewHolder, dX, dY, actionState, isCurrentlyActive);
-
-        if (actionState == ItemTouchHelper.ACTION_STATE_SWIPE) {
-
+        if (viewHolder instanceof TodoAdapter.ViewHolder && actionState == ItemTouchHelper.ACTION_STATE_SWIPE) {
+            super.onChildDraw(c, recyclerView, viewHolder, dX, dY, actionState, isCurrentlyActive);
             Drawable d = ContextCompat.getDrawable(context, R.drawable.delete_icon);
-            Paint rect = new Paint(Paint.ANTI_ALIAS_FLAG);
-            rect.setColor(Color.BLACK);
-            rect.setAlpha(50);
-
             View item = viewHolder.itemView;// item currently getting scrolled
             int left = item.getLeft(), right = item.getRight(), top = item.getTop(), bottom = item.getBottom(), height = item.getHeight();
+            Paint rect = new Paint(Paint.ANTI_ALIAS_FLAG);
+            rect.setColor(Color.RED);
             c.drawRect(left, top, left + ((int) dX) + 20, bottom, rect);
 
 
@@ -60,17 +60,7 @@ public class TodoTouchHelperCallback extends ItemTouchHelper.SimpleCallback {
             int iconWidth = d.getIntrinsicWidth();
             int iconCentre = iconLeft + (iconWidth / 2);
 
-            if (left + ((int) dX) > iconWidth + (height / 2) - 25) {
-
-                Paint circle = new Paint(Paint.ANTI_ALIAS_FLAG);
-                circle.setColor(Color.RED);
-
-                int radius = (height / 2) - 25;
-
-                if (iconCentre + radius > left + ((int) dX))
-                    radius = left + ((int) dX) - iconCentre - 5;
-
-                c.drawCircle(iconLeft + (iconWidth / 2), iconTop + (iconHeight / 2), radius, circle);
+            if (left + ((int) dX) > left +iconWidth +iconMargin+20) {
                 d.setBounds(iconLeft, iconTop, iconRight, iconBottom);
                 d.draw(c);
             }
@@ -79,17 +69,18 @@ public class TodoTouchHelperCallback extends ItemTouchHelper.SimpleCallback {
 
     @Override
     public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int direction) {
-        if (direction == ItemTouchHelper.RIGHT) {
+        if (viewHolder instanceof TodoAdapter.ViewHolder && direction == ItemTouchHelper.RIGHT) {
 
-            int position = viewHolder.getAdapterPosition();
-            Todo item = adapter.getCurrentList().get(position);
+            int position = viewHolder.getAbsoluteAdapterPosition();
+            Todo item = (Todo) adapter.getCurrentList().get(position);
             repository.delete(item);
-
             Snackbar snackbar = Snackbar.make(parent, "Task deleted", Snackbar.LENGTH_LONG);
-            snackbar.setAction("Undo", new View.OnClickListener() {
+            snackbar.setAction("Undo", v -> repository.insert(item));
+            snackbar.addCallback(new BaseTransientBottomBar.BaseCallback<Snackbar>() {
                 @Override
-                public void onClick(View v) {
-                    repository.insert(item);
+                public void onDismissed(Snackbar transientBottomBar, int event) {
+                    super.onDismissed(transientBottomBar, event);
+                    NotificationHelper.cancel(context,item.getNid());
                 }
             });
             snackbar.show();
